requirements.txt
WeasyPrint==51
Django==3.1.1
django_elasticsearch_dsl==7.1.4
django_filter==2.3.0
Pillow==8.1.2
